convert module-loader.test to TypeScript
